Split log-ingest object-storage into api and core layers
The S3 client and bucket readiness probe are infrastructure adapters,
not domain behavior. Move them under src/api/ so the layer signals
intent, and lift logObjectKey + LogObjectKeyParams into
core/entities/log-object.ts where the key layout belongs.
